+++ lib/rivertion.dart
@@ -1,6 +1,6 @@
 library;
 
 export 'src/source.dart';
-export 'src/source_notifier.dart';
+export 'src/source_controller.dart';
 export 'src/source_widgets.dart';
 export 'src/sources/listenable_source.dart';