+++ lib/src/internals.dart
@@ -1,3 +1,4 @@
 export '../rivertion.dart';
+export 'internals/source_listenable_extension.dart';
 export 'internals/source_stateful_element_mixin.dart';
 export 'internals/source_subscriptions.dart';