+++ lib/src/source.dart
@@ -8,30 +8,31 @@ typedef SourceListener<T> = void Function(T previous, T state);
 
 typedef SourceImmediatelyListener<T> = void Function(T? previous, T state);
 
-abstract interface class SourceProvider<T> {
-  @internal
-  Source<T> get source;
+sealed class Source<T> {}
+
+mixin SourceContainer<T> implements Source<T> {
+  SourceListenable<T> get source;
 }
 
 @immutable
-abstract base class Source<T> implements SourceProvider<T> {
-  @override
-  Source<T> get source => this;
-
-  SourceSubscription<T> listen(SourceListener<T> onChange);
+abstract base class SourceListenable<T> implements Source<T> {
+  SourceSubscription<T> listen(SourceListener<T> listener);
 
   SourceSubscription<T> listenImmediately(SourceImmediatelyListener<T> listener) {
     final subscription = listen(listener);
     Zone.current.runBinaryGuarded(listener, null, subscription.read());
     return subscription;
   }
+}
 
-  Source<T> where(bool Function(T previous, T next) condition) => SourceFilter(this, condition);
+extension SourceExtensions<T> on Source<T> {
+  SourceListenable<R> select<R>(R Function(T state) selector) => SourceSelector(this, selector);
 
-  Source<R> select<R>(R Function(T state) selector) => SourceSelector(this, selector);
-
-  Source<R> selectWith<R, A>(A arg, R Function(A arg, T state) selector) =>
+  SourceListenable<R> selectWith<R, A>(A arg, R Function(A arg, T state) selector) =>
       SourceArgSelector(this, arg, selector);
+
+  SourceListenable<T> where(bool Function(T previous, T next) condition) =>
+      SourceFilter(this, condition);
 }
 
 abstract base class SourceSubscription<T> {