+++ tool/generate_templates.dart
@@ -16,7 +16,7 @@ ${templates.map((template) {
     ${content.contains('\$') ? 'r' : ''}'''
 $content''',
   );
-  """;
+""";
   }).join('\n')}
 }
     ''');