+++ tool/templates/reactive_forms_sources.dart
@@ -8,25 +8,26 @@ import 'package:reactive_forms/reactive_forms.dart';
 import 'package:rivertion/src/internals.dart';
 
 extension AbstractControlStateSource<V> on AbstractControl<V> {
-  Source<AbstractControlState<V?>> get source => _AbstractControlStateSource(this);
+  SourceListenable<AbstractControlState<V?>> get source => _AbstractControlStateSource(this);
 
-  Source<R> select<R>(R Function(AbstractControl<V> control) selector) =>
+  SourceListenable<R> select<R>(R Function(AbstractControl<V> control) selector) =>
       _FormControlSource(this).select(selector);
 }
 
 extension ControlStateSource<C extends AbstractControl<Object?>> on C {
-  Source<R> select<R>(R Function(C control) selector) => _FormControlSource(this).select(selector);
+  SourceListenable<R> select<R>(R Function(C control) selector) =>
+      _FormControlSource(this).select(selector);
 }
 
-extension AbstractControlStateSourceExtensions<V> on Source<AbstractControlState<V>> {
-  Source<bool> get hasValue => select(_hasValue);
-  Source<V?> get value => select(_value);
-  Source<bool> get pristine => select(_pristine);
-  Source<bool> get dirty => select(_dirty);
-  Source<bool> get touched => select(_touched);
-  Source<ControlStatus> get status => select(_status);
-  Source<MapEntry<String, Object>?> get error => select(_error);
-  Source<bool> get isEmpty => select(_isEmpty);
+extension AbstractControlStateSourceExtensions<V> on SourceListenable<AbstractControlState<V>> {
+  SourceListenable<bool> get hasValue => select(_hasValue);
+  SourceListenable<V?> get value => select(_value);
+  SourceListenable<bool> get pristine => select(_pristine);
+  SourceListenable<bool> get dirty => select(_dirty);
+  SourceListenable<bool> get touched => select(_touched);
+  SourceListenable<ControlStatus> get status => select(_status);
+  SourceListenable<MapEntry<String, Object>?> get error => select(_error);
+  SourceListenable<bool> get isEmpty => select(_isEmpty);
 
   static bool _hasValue<V>(AbstractControlState<V> state) => state.value != null;
   static V? _value<V>(AbstractControlState<V> state) => state.value;
@@ -44,10 +45,10 @@ extension AbstractControlStateSourceExtensions<V> on Source<AbstractControlState
   }
 }
 
-extension ControlStatusSourceExtensions on Source<ControlStatus> {
-  Source<bool> get enabled => select(_enabled);
-  Source<bool> get disabled => select(_disabled);
-  Source<bool> get valid => select(_valid);
+extension ControlStatusSourceExtensions on SourceListenable<ControlStatus> {
+  SourceListenable<bool> get enabled => select(_enabled);
+  SourceListenable<bool> get disabled => select(_disabled);
+  SourceListenable<bool> get valid => select(_valid);
 
   static bool _enabled(ControlStatus status) => status != ControlStatus.disabled;
   static bool _disabled(ControlStatus status) => status == ControlStatus.disabled;
@@ -55,11 +56,11 @@ extension ControlStatusSourceExtensions on Source<ControlStatus> {
 }
 
 extension FormControlStateSource<V> on FormControl<V> {
-  Source<FormControlState<V>> get source => _FormControlStateSource(this);
+  SourceListenable<FormControlState<V>> get source => _FormControlStateSource(this);
 }
 
-extension FormControlStateSourceExtensions<V> on Source<FormControlState<V?>> {
-  Source<bool> get hasFocus => select(_hasFocus);
+extension FormControlStateSourceExtensions<V> on SourceListenable<FormControlState<V?>> {
+  SourceListenable<bool> get hasFocus => select(_hasFocus);
 
   static bool _hasFocus<V>(FormControlState<V?> state) => state.hasFocus;
 }
@@ -93,7 +94,7 @@ class AbstractControlState<V> {
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      other is AbstractControlState &&
+      other is AbstractControlState<V> &&
           runtimeType == other.runtimeType &&
           value == other.value &&
           pristine == other.pristine &&
@@ -174,7 +175,7 @@ abstract base class _AbstractControlStateSourceBase<
   TControl extends AbstractControl<Object?>,
   TState extends AbstractControlState<Object?>
 >
-    extends Source<TState> {
+    extends SourceListenable<TState> {
   final TControl control;
 
   _AbstractControlStateSourceBase(this.control);
@@ -205,7 +206,7 @@ abstract base class _AbstractControlStateSourceBase<
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      other is _AbstractControlStateSourceBase &&
+      other is _AbstractControlStateSourceBase<TControl, TState> &&
           runtimeType == other.runtimeType &&
           control == other.control;
 
@@ -213,7 +214,8 @@ abstract base class _AbstractControlStateSourceBase<
   int get hashCode => control.hashCode;
 }
 
-final class _FormControlSource<TControl extends AbstractControl<Object?>> extends Source<TControl> {
+final class _FormControlSource<TControl extends AbstractControl<Object?>>
+    extends SourceListenable<TControl> {
   final TControl control;
 
   _FormControlSource(this.control);
@@ -236,7 +238,7 @@ final class _FormControlSource<TControl extends AbstractControl<Object?>> extend
   @override
   bool operator ==(Object other) =>
       identical(this, other) ||
-      other is _AbstractControlStateSourceBase &&
+      other is _FormControlSource<TControl> &&
           runtimeType == other.runtimeType &&
           control == other.control;